refactor: Consolidate SheetMusicMetrics into SheetMusicLayout
- Merge SheetMusicMetrics functionality into SheetMusicLayout class
- Remove sheet_music_metrics.dart and related mock/test files
- Update simple_sheet_music.dart to use consolidated SheetMusicLayout
- Rename test file from sheet_music_metrics_test to sheet_music_layout_test
- Add consolidation plan document

This follows the same pattern as PR #43 which unified other Metrics classes.
