+++ test/sheet_music_layout_test.dart
@@ -1,31 +1,90 @@
-import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:simple_sheet_music/src/music_objects/clef/clef_type.dart';
+import 'package:simple_sheet_music/src/music_objects/key_signature/keysignature_type.dart';
 import 'package:simple_sheet_music/src/sheet_music_layout.dart';
+import 'package:flutter/material.dart';
 
-import 'mock/mock_sheet_music_metrics.dart';
+import 'mock/mocks.dart';
 
 void main() {
-  test(
-      'SheetMusicLayout.canvasScale should be the minimum scale factor to fit the sheet music in the widget',
-      () {
+  test('SheetMusicLayout should calculate staffRenderers correctly', () {
     // Arrange
-    const maximumStaffWidth = 100.0;
-    final metrics = MockSheetMusicMetrics(
-      maximumStaffWidth: maximumStaffWidth,
-      staffsHeightSum: 1,
-    );
-    const widgetWidth = 100.0;
-    final layout = SheetMusicLayout(
-      metrics,
-      Colors.black,
-      widgetHeight: 100,
-      widgetWidth: widgetWidth,
+    final measures = [
+      MockMeasure(),
+      MockMeasure(),
+      MockMeasure(),
+    ];
+    final sheetMusicLayout = SheetMusicLayout(
+      musicalSymbols: measures,
+      initialClefType: ClefType.treble,
+      initialKeySignatureType: KeySignatureType.cMajor,
+      metadata: MockGlyphMetadata(),
+      paths: MockGlyphPath(),
+      lineColor: Colors.black,
+      widgetWidth: 400,
+      widgetHeight: 400,
     );
 
     // Act
-    final canvasScale = layout.canvasScale;
+    final staffRenderers = sheetMusicLayout.staffRenderers;
 
     // Assert
-    expect(canvasScale, maximumStaffWidth / widgetWidth);
+    expect(staffRenderers.length, equals(1));
+    expect(staffRenderers.first.measureRenderers.length, equals(3));
+  });
+  test(
+      'SheetMusicLayout should calculate staffRenderers correctly even if there is a new line measure',
+      () {
+    // Arrange
+    final measures = [
+      MockMeasure(),
+      MockMeasure(),
+      MockMeasure(isNewLine: true),
+    ];
+    final sheetMusicLayout = SheetMusicLayout(
+      musicalSymbols: measures,
+      initialClefType: ClefType.treble,
+      initialKeySignatureType: KeySignatureType.cMajor,
+      metadata: MockGlyphMetadata(),
+      paths: MockGlyphPath(),
+      lineColor: Colors.black,
+      widgetWidth: 400,
+      widgetHeight: 400,
+    );
+
+    // Act
+    final staffRenderers = sheetMusicLayout.staffRenderers;
+
+    // Assert
+    expect(staffRenderers.length, equals(2));
+    expect(staffRenderers[0].measureRenderers.length, equals(2));
+    expect(staffRenderers[1].measureRenderers.length, equals(1));
+  });
+  test(
+      'SheetMusicLayout should calculate staffRenderers correctly even if first measure is a new line measure',
+      () {
+    // Arrange
+    final measures = [
+      MockMeasure(isNewLine: true),
+      MockMeasure(),
+      MockMeasure(),
+    ];
+    final sheetMusicLayout = SheetMusicLayout(
+      musicalSymbols: measures,
+      initialClefType: ClefType.treble,
+      initialKeySignatureType: KeySignatureType.cMajor,
+      metadata: MockGlyphMetadata(),
+      paths: MockGlyphPath(),
+      lineColor: Colors.black,
+      widgetWidth: 400,
+      widgetHeight: 400,
+    );
+
+    // Act
+    final staffRenderers = sheetMusicLayout.staffRenderers;
+
+    // Assert
+    expect(staffRenderers.length, equals(1));
+    expect(staffRenderers.first.measureRenderers.length, equals(3));
   });
 }
